Start moving test_replace.py
