+++ python/cudf/cudf/tests/dataframe/methods/test_where.py
@@ -1,0 +1,81 @@
+# Copyright (c) 2025, NVIDIA CORPORATION.
+
+
+import numpy as np
+import pandas as pd
+import pytest
+
+import cudf
+from cudf.testing import assert_eq
+
+
+@pytest.mark.parametrize("fill_value", [[888, 999]])
+def test_dataframe_with_nulls_where_with_scalars(fill_value):
+    pdf = pd.DataFrame(
+        {
+            "A": [-1, 2, -3, None, 5, 6, -7, 0],
+            "B": [4, -2, 3, None, 7, 6, 8, 0],
+        }
+    )
+    gdf = cudf.from_pandas(pdf)
+
+    expect = pdf.where(pdf % 3 == 0, fill_value)
+    got = gdf.where(gdf % 3 == 0, fill_value)
+
+    assert_eq(expect, got)
+
+
+def test_dataframe_with_different_types():
+    # Testing for int and float
+    pdf = pd.DataFrame(
+        {"A": [111, 22, 31, 410, 56], "B": [-10.12, 121.2, 45.7, 98.4, 87.6]}
+    )
+    gdf = cudf.from_pandas(pdf)
+    expect = pdf.where(pdf > 50, -pdf)
+    got = gdf.where(gdf > 50, -gdf)
+
+    assert_eq(expect, got)
+
+    # Testing for string
+    pdf = pd.DataFrame({"A": ["a", "bc", "cde", "fghi"]})
+    gdf = cudf.from_pandas(pdf)
+    pdf_mask = pd.DataFrame({"A": [True, False, True, False]})
+    gdf_mask = cudf.from_pandas(pdf_mask)
+    expect = pdf.where(pdf_mask, ["cudf"])
+    got = gdf.where(gdf_mask, ["cudf"])
+
+    assert_eq(expect, got)
+
+    # Testing for categoriacal
+    pdf = pd.DataFrame({"A": ["a", "b", "b", "c"]})
+    pdf["A"] = pdf["A"].astype("category")
+    gdf = cudf.from_pandas(pdf)
+    expect = pdf.where(pdf_mask, "c")
+    got = gdf.where(gdf_mask, ["c"])
+
+    assert_eq(expect, got)
+
+
+def test_dataframe_where_with_different_options():
+    pdf = pd.DataFrame({"A": [1, 2, 3], "B": [3, 4, 5]})
+    gdf = cudf.from_pandas(pdf)
+
+    # numpy array
+    boolean_mask = np.array([[False, True], [True, False], [False, True]])
+
+    expect = pdf.where(boolean_mask, -pdf)
+    got = gdf.where(boolean_mask, -gdf)
+
+    assert_eq(expect, got)
+
+    # with single scalar
+    expect = pdf.where(boolean_mask, 8)
+    got = gdf.where(boolean_mask, 8)
+
+    assert_eq(expect, got)
+
+    # with multi scalar
+    expect = pdf.where(boolean_mask, [8, 9])
+    got = gdf.where(boolean_mask, [8, 9])
+
+    assert_eq(expect, got)
